repair warning

+++ services/param/linux/param_service.c
@@ -165,7 +165,7 @@ static int SystemUpdateConstParamter(const char *name, const char *value, const 
     int ret = CheckParameterSet(name, value, srcLabel, &ctrlService);
     PARAM_CHECK(ret == 0, return ret, "Forbid to set parameter %s", name);
     PARAM_CHECK(IS_READY_ONLY(name), return PARAM_CODE_INVALID_NAME, "only update read only param: %s", name);
-    if ((ctrlService & PARAM_CTRL_SERVICE) != PARAM_CTRL_SERVICE) { // ctrl param
+    if (((unsigned int)ctrlService & PARAM_CTRL_SERVICE) != PARAM_CTRL_SERVICE) { // ctrl param
         uint32_t dataIndex = 0;
         ret = WriteParam(name, value, &dataIndex, LOAD_PARAM_UPDATE_CONST);
         PARAM_CHECK(ret == 0, return ret, "Failed to update const param %d name %s %s", ret, name, value);
